pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "plaudsync"
version = "0.0.1"
description = "Periodic sync of Plaud AI recordings to local disk, categorized via M365 / regex / LLM waterfall."
readme = "SPEC.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "tomis", email = "tomis@kvados.cz" }]

dependencies = [
    # HTTP clients
    "httpx>=0.27",
    "requests>=2.32",
    # Microsoft Graph
    "msal>=1.30",
    # LLM provider
    "anthropic>=0.40",
    # Observability
    "loguru>=0.7.2",
    "sentry-sdk>=2.10",
    # Env loading
    "python-dotenv>=1.0",
    # Config
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    # Test framework
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    # Integration tests — HTTP replay
    "vcrpy>=8.0",
    "pytest-recording>=0.13",
    # LLM evals
    "deepeval>=1.5",
    # Security SAST
    "bandit>=1.7",
    # Packaging (optional, activate when distributing)
    # "pyinstaller>=6.10",
]

[project.scripts]
plaudsync = "plaudsync.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
# pytest-recording: block real network by default; override per-test with @pytest.mark.vcr()
addopts = [
    "-ra",
    "--strict-markers",
    "--block-network",
]
markers = [
    "vcr: integration test using VCR cassette (pytest-recording)",
    "eval: LLM classifier eval against golden set (DeepEval)",
    "slow: test > 1s wall clock",
]
# VCR default cassette location
# Individual tests can override via @pytest.mark.vcr(cassette_library_dir=...)

[tool.bandit]
exclude_dirs = ["tests", ".venv", "build", "dist"]
skips = []  # add rule IDs here if false positives warrant it
